fix(chat): bundle LLM provider logos instead of loading them from models.dev
Model/provider icons were loaded at render time from the models.dev CDN,
so any environment that blocks third-party requests (mobile content
blockers, restrictive networks, air-gapped deployments) rendered a broken
empty box in the chat composer and everywhere else the logos appear.

Bundle the SVGs under public/model-logos and serve them from our own
origin. Consolidates the three duplicated provider->logo maps
(chat model-selector, llm-model-select, optimization-rules) into
lib/provider-logos.ts; the optimization-rules provider select previously
skipped the mapping entirely and showed a generic fallback icon for
gemini/bedrock/ollama/kimi, which the shared helper also fixes. A test
pins that every mapped provider has a bundled logo file.
