+++ platform/frontend/src/components/ai-elements/model-selector.tsx
@@ -175,7 +175,7 @@ export const ModelSelectorLogo = ({
     alt={`${provider} logo`}
     className={cn("size-4 dark:invert", className)}
     height={12}
-    src={`https://models.dev/logos/${provider}.svg`}
+    src={`/model-logos/${provider}.svg`}
     width={12}
   />
 );
