+++ platform/frontend/src/components/chat/model-selector.tsx
@@ -59,6 +59,7 @@ import {
   type ModelCapabilities,
   useLlmModelsByProvider,
 } from "@/lib/llm-models.query";
+import { providerToLogoProvider } from "@/lib/provider-logos";
 import { cn, formatContextLength } from "@/lib/utils";
 
 /** Modalities that can be filtered (excludes "text" since all models support it) */
@@ -116,35 +117,6 @@ interface ModelSelectorProps {
    */
   fallbackModelName?: string;
 }
-
-/** Map our provider names to logo provider names
- * models.dev provider IDs
- * see https://github.com/anomalyco/models.dev/tree/dev/providers
- * */
-export const providerToLogoProvider: Record<SupportedProvider, string> = {
-  openai: "openai",
-  anthropic: "anthropic",
-  gemini: "google",
-  bedrock: "amazon-bedrock",
-  cerebras: "cerebras",
-  cohere: "cohere",
-  mistral: "mistral",
-  perplexity: "perplexity",
-  groq: "groq",
-  xai: "xai",
-  openrouter: "openrouter",
-  vllm: "vllm",
-  ollama: "ollama-cloud", // models.dev uses ollama-cloud for the Ollama provider
-  "ollama-native": "ollama-cloud",
-  zhipuai: "zhipuai",
-  deepseek: "deepseek",
-  minimax: "minimax",
-  kimi: "moonshotai",
-  azure: "azure",
-  "github-copilot": "github-copilot",
-  "microsoft-365-copilot": "microsoft-365-copilot",
-  archestra: "archestra",
-};
 
 /**
  * Creates a unique value for a model that includes the provider.
